goforawalk/requirements.txt
django-appconf==1.0.5
django-compressor==4.1
django-push-notifications==3.0.0
django-pwa==1.0.10
django-sass==1.1.0
django-sass-processor==1.2.2
django-simple-pwa==2.2.1
django-webpush==0.3.4
djangorestframework==3.14.0
fcm-django==1.0.12
firebase-admin==5.4.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
psycopg2==2.9.5







whitenoise==6.2.0
